Move des from FileFormats to Crypto
