+++ src/openboardview/FileFormats/XZZPCBFile.cpp
@@ -13,7 +13,7 @@
 #include <unordered_map>
 #include <vector>
 
-#include "des.h"
+#include "Crypto/des.h"
 #include "utils.h"
 
 /*
